add user - item relations for item actions

+++ WishList/Controllers/ItemController.cs
@@ -1,23 +1,33 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
-
-            return View("Index", model);
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            if (user != null)
+            {
+                var model = _context.Items.Where(x => x.User.Id == user.Id).ToList();
+                return View("Index", model);
+            }
+            
         }
 
         [HttpGet]
@@ -29,6 +39,11 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            if(user != null)
+            {
+                item.User = user;
+            }
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -37,6 +52,10 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            if (item.User != user)
+                return Unauthorized();
+
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
